COPY Extended to work for ALL File Types
COPY NOT DELETING CHUNKS(NO WAY TO cancel copy operation)[FIXED]

Add Stipulation for either one copy or to Iterate the copy

PROBLEM: Automatically set Logged in Flag to 0 on App Exit

+++ soft400051_hj_local/src/main/java/com/mycompany/soft400051_hj_local/UserProfileController.java
@@ -215,14 +215,16 @@ public class UserProfileController extends FileMethods implements Initializable 
     
     public void DeleteBtnClicked()
     {
-        Alert btnAlert = new Alert(AlertType.WARNING);
+        Alert btnAlert = new Alert(AlertType.CONFIRMATION);
         btnAlert.contentTextProperty().setValue("Are you Sure you want to Delete this File?");
+        
         Optional<ButtonType> result = btnAlert.showAndWait();
-        if(result.isPresent())
+        if(result.isPresent() && result.get().getText().equals("OK"))
         {
             DeleteFile(tableFiles.getSelectionModel().getSelectedItem());
             refreshGrid();
-        }
+        }  
+        
     }
     
     public void RecycleBtnClicked()
